… core/trino-main/src/main/java/io/trino/likematcher/Pattern.java
… core/trino-main/src/main/java/io/trino/sql/planner/ExpressionInterpreter.java
… core/trino-main/src/main/java/io/trino/type/LikeFunctions.java
… core/trino-main/src/main/java/io/trino/type/LikePatternType.java
… core/trino-main/src/test/java/io/trino/likematcher/TestLikeMatcher.java
… core/trino-main/src/test/java/io/trino/operator/scalar/BenchmarkLike.java
… core/trino-main/src/test/java/io/trino/operator/scalar/TestJoniRegexpFunctions.java
… core/trino-main/src/test/java/io/trino/operator/scalar/TestLikeFunctions.java
… core/trino-main/src/test/java/io/trino/sql/gen/TestExpressionCompiler.java
… core/trino-main/src/test/java/io/trino/sql/planner/TestLiteralEncoder.java
… core/trino-main/src/test/java/io/trino/type/TestLikePatternType.java