Implements a dedicated NFA/DFA-based matcher for LIKE expressions.

Before:
```
  (pattern)  Mode  Cnt    Score    Error  Units
          %  avgt   90  249.201 ±  1.761  ns/op
         _%  avgt   90  290.866 ±  0.947  ns/op
         %_  avgt   90  299.377 ±  2.079  ns/op
       abc%  avgt   90  244.724 ±  2.004  ns/op
       %abc  avgt   90  139.165 ±  7.407  ns/op
      _____  avgt   90   25.508 ±  0.161  ns/op
abc%def%ghi  avgt   90  209.570 ± 11.658  ns/op
  %abc%def%  avgt   90  447.848 ±  0.739  ns/op
```

After:
```
          %  avgt   90    3.989 ±  0.012  ns/op
         _%  avgt   90    3.969 ±  0.014  ns/op
         %_  avgt   90    3.958 ±  0.013  ns/op
       abc%  avgt   90    5.501 ±  0.019  ns/op
       %abc  avgt   90    5.861 ±  0.015  ns/op
      _____  avgt   90    7.096 ±  0.028  ns/op
abc%def%ghi  avgt   90   84.293 ±  0.623  ns/op
  %abc%def%  avgt   90  152.031 ±  0.247  ns/op
```

## Documentation

(x) No documentation is needed.

## Release notes

(x) Release notes entries required with the following suggested text:

```markdown
# General
* Improve performance of `LIKE` expressions. ({issue}`13479`)
```