@martint 2022-08-22
> Looking through how this PR works, I think there's something of an extra step here in the construction of the DFA. It looks like the DFA class is only used as a transitory state between NFA and DenseDfaMatcher, the latter of which actually functions as a DFA.

The DFA class is an abstract description of the DFA (useful for understanding the states and transitions). The one used by DfaMatcher is a physical version designed for efficient matching.

> Either way, I think the part of LikeMatcher.compile that essentially compiles a DFA should be moved to its own method on the current DFA class.

You make a good point. That bit of code should actually go in DenseDfaMatcher, which should provide a method for constructing a instance of that class from the abstract representation of the DFA.